requirements.txt
Flask>=2.0.3
Flask-SQLAlchemy>=2.5.1
psycopg2>=2.9.3
argon2-cffi>=21.3.0